backend/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.34.0
sqlalchemy>=2.0.40
alembic>=1.15.0
psycopg[binary]>=3.2.0
httpx>=0.28.0
pydantic>=2.11.0
pydantic-settings>=2.9.0
python-jose[cryptography]>=3.4.0
passlib[bcrypt]>=1.7.4
python-multipart>=0.0.20
apscheduler>=3.11.0
